+++ include/Windowing/WindowActions/ExecutableCompleteAggregationTriggerAction.hpp
@@ -64,7 +64,7 @@ class ExecutableCompleteAggregationTriggerAction : public BaseExecutableWindowAc
         windowTupleLayout = createRowLayout(this->windowSchema);
     }
 
-    bool doAction(StateVariable<KeyType, WindowSliceStore<PartialAggregateType>*>* windowStateVariable, size_t currentWatermark, size_t lastWatermark) {
+    bool doAction(StateVariable<KeyType, WindowSliceStore<PartialAggregateType>*>* windowStateVariable, uint64_t currentWatermark, uint64_t lastWatermark) {
         NES_DEBUG("ExecutableCompleteAggregationTriggerAction: doAction for currentWatermark=" << currentWatermark << " lastWatermark=" << lastWatermark);
 
         auto tupleBuffer = this->bufferManager->getBufferBlocking();
@@ -105,8 +105,8 @@ class ExecutableCompleteAggregationTriggerAction : public BaseExecutableWindowAc
                           WindowSliceStore<PartialAggregateType>* store,
                           LogicalWindowDefinitionPtr windowDefinition,
                           TupleBuffer& tupleBuffer,
-                          size_t currentWatermark,
-                          size_t lastWatermark) {
+                          uint64_t currentWatermark,
+                          uint64_t lastWatermark) {
 
         // For event time we use the maximal records ts as watermark.
         // For processing time we use the current wall clock as watermark.
